dialog action

+++ app/lib/widgets/vote.dart
@@ -1,11 +1,11 @@
 import 'dart:async';
 
+import 'package:consensor/widgets/vote_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 import 'package:consensor/models/vote.dart';
-import 'package:consensor/routes/vote.dart';
 import 'package:consensor/services/vote.dart';
 
 class VoteWidget extends StatefulWidget {
@@ -20,15 +20,18 @@ class VoteWidget extends StatefulWidget {
 
 class _VoteWidgetState extends State<VoteWidget> {
   VoteService _voteSvc = new VoteService();
-  List<Vote> _items;
+  List<Vote> _votes;
 
   StreamSubscription<QuerySnapshot> _voteSub;
 
+  bool _isLoaded;
+
   @override
   void initState() {
+    _isLoaded = false;
     super.initState();
 
-    _items = new List();
+    _votes = new List();
 
     _voteSub?.cancel();
     _voteSub = _voteSvc.getVoteList().listen((QuerySnapshot snapshot) {
@@ -37,47 +40,58 @@ class _VoteWidgetState extends State<VoteWidget> {
           .toList();
 
       setState(() {
-        this._items = votes;
+        this._votes = votes;
+        this._isLoaded = true;
       });
     });
   }
 
   @override
   Widget build(BuildContext context) {
-    if (_items.length > 0) {
-      return Padding(
-          padding: const EdgeInsets.symmetric(vertical: 5.0),
-          child: Center(
-            child: ListView.builder(
-                itemCount: _items.length,
-                padding: const EdgeInsets.all(15.0),
-                itemBuilder: (context, position) {
-                  return Column(
-                    children: <Widget>[
-                      Card(
-                        child: Column(
-                          mainAxisSize: MainAxisSize.min,
-                          children: <Widget>[
-                            ListTile(
-                                leading: Icon(Icons.done),
-                                title: Text(
-                                  '${_items[position].topic}',
-                                  style: TextStyle(fontSize: 22.0),
-                                ),
-                                subtitle: Text(
-                                    'expires on ${_items[position].expiresOn()}'),
-                                trailing: IconButton(
-                                    icon:
-                                        const Icon(Icons.remove_circle_outline),
-                                    onPressed: () => _deleteVote(
-                                        context, _items[position], position)),
-                                onTap: () {/* set topic */}),
-                            ButtonTheme.bar(
-                              child: ButtonBar(
+    if (_isLoaded) {
+      if (_votes.length > 0) {
+        return Padding(
+            padding: const EdgeInsets.symmetric(vertical: 5.0),
+            child: Center(
+              child: ListView.builder(
+                  itemCount: _votes.length,
+                  padding: const EdgeInsets.all(15.0),
+                  itemBuilder: (context, position) {
+                    return Column(
+                      children: <Widget>[
+                        Card(
+                          child: Column(
+                            mainAxisSize: MainAxisSize.min,
+                            children: <Widget>[
+                              ListTile(
+                                  leading: Icon(Icons.done),
+                                  title: Text(
+                                    '${_votes[position].topic}',
+                                    style: TextStyle(fontSize: 22.0),
+                                  ),
+                                  subtitle: Text(
+                                      'expires on ${_votes[position].expiresOn()}'),
+                                  trailing: IconButton(
+                                      icon: const Icon(
+                                          Icons.remove_circle_outline),
+                                      onPressed: () => _deleteVote(
+                                          context, _votes[position], position)),
+                                  onTap: () {/* set topic */}),
+                              ButtonBar(
                                 children: <Widget>[
                                   FlatButton(
                                     child: Text('OPTIONS'),
-                                    onPressed: () {/* ... */},
+                                    onPressed: () {
+                                      showDialog(
+                                          context: context,
+                                          builder: (BuildContext context) =>
+                                              MyDialog(
+                                                title: "Success",
+                                                description:
+                                                    "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
+                                                buttonText: "Okay",
+                                              ));
+                                    },
                                   ),
                                   FlatButton(
                                     child: Text('GROUP'),
@@ -88,42 +102,61 @@ class _VoteWidgetState extends State<VoteWidget> {
                                     onPressed: () {/* ... */},
                                   ),
                                   FlatButton(
-                                    child: Text(_items[position].enabled
+                                    child: Text(_votes[position].enabled
                                         ? 'STOP'
                                         : 'START'),
                                     onPressed: () {/* ... */},
                                   ),
                                 ],
                               ),
-                            ),
-                          ],
+                            ],
+                          ),
                         ),
-                      ),
-                      Padding(padding: EdgeInsets.all(5.0)),
-                    ],
-                  );
-                }),
-          ));
+                        Padding(padding: EdgeInsets.all(5.0)),
+                      ],
+                    );
+                  }),
+            ));
+      } else {
+        return Padding(
+            padding: const EdgeInsets.symmetric(vertical: 5.0),
+            child: Center(child: Text('Create a vote to get started')));
+      }
     } else {
-      return Padding(
-          padding: const EdgeInsets.symmetric(vertical: 5.0),
-          child: Center(child: Text('Create a vote to get started')));
+      return Center(
+          child: SizedBox(
+        child: CircularProgressIndicator(),
+        height: 100.0,
+        width: 100.0,
+      ));
     }
   }
 
   void _deleteVote(BuildContext context, Vote vote, int position) async {
-    _voteSvc.deleteVote(vote.id).then((votes) {
-      setState(() {
-        _items.removeAt(position);
-      });
-    });
-  }
-
-  void _navigateToVote(BuildContext context, Vote vote) async {
-    await Navigator.push(
-      context,
-      MaterialPageRoute(
-          builder: (context) => VotePage(vote, widget.user, widget.onWaiting)),
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: new Text("Delete Confirmation"),
+          content: new Text("Are you sure you want to delete this vote?"),
+          actions: <Widget>[
+            new FlatButton(
+              child: new Text("Close"),
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+            ),
+            new FlatButton(
+              child: new Text("Confirm"),
+              onPressed: () {
+                _voteSvc.deleteVote(vote.id).then((votes) {
+                  Navigator.of(context).pop();
+                });
+              },
+            )
+          ],
+        );
+      },
     );
   }
